+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `IO#safe`, mapping an `IO<A>` to an `IO<Either<Throwable, A>>` that will never throw
 - `IO#ensuring`, like `finally` semantics for `IO`s
 - `IO#throwing`, for producing an `IO<A>` that will throw a given `Throwable` when executed
+- `Bracket`, for bracketing an `IO` operation with a mapping operation and a cleanup operation 
 
 ## [3.3.0] - 2019-02-18
 ### Added